Rewrite Identity in Rust (and GTK 4)
Who said you can't RiiR a project that was already in Rust?

Jokes aside, this is a GTK 4 port which is more of a complete rewrite
while referencing the existing code. It uses GTK 4 and libadwaita and
the most up-to-date gtk-rs idioms.

The stack view switcher has been replaced with libadwaita tabs with the
newly added ability to drag a tab into a new window and drag tabs
between windows.

gtksink has been replaced with gtk4paintablesink from gst-plugin-gtk4.
As a side effect, this enables hardware-accelerated video decoding.

Since Identity is a multimedia application, the default style has been
set to prefer-dark.

All .ui files have been ported to blueprint, following my successful use
of it in Video Trimmer.