Add initial tiled display mode implementation
The new PageGrid widget, backed by a Box, is responsible for the tiled
display. It mirrors some of the API of AdwTabView that Identity uses.

When transitioning between display modes, all pages are removed from one
widget and added to the other. This is a bit ugly and requires some
hacks, but it's the easiest with the current code architecture.

A potential refactor idea for the future is to decouple a
playbin-containing object from a view widget (currently both are Page's
responsibility). This will allow having two or more views of the same
playbin, and will allow having views in both AdwTabView and PageGrid at
the same time, removing the need for some of the hacks. However, then
you would also need to be careful not to leave the source attached to
one window with a view of that source in another window.

Anyway, this tiled display mostly works, but is still missing some
things like the context menu. It also has no animations, partially
because of the problem above with coupled-together Page, and partially
because I'm already running out of motivation and want to release
something working, maybe if not entirely pretty looking.

Some things (current scale, media properties, Ctrl-W) rely on having a
single "selected" Page. For this, I added focus rings to Page and
"selection" to PageGrid, visually indicated in a mostly expected way.
The only way to change the selected Page is currently to focus it with
Tab/Shift-Tab and press Space/Enter. I'll think about how to integrate
the left click into this later (it's not trivial as it's also involved
in panning, and I also wanted to add general DnD when zoomed out, like
in Loupe).