player: Join processing thread on dispose
Otherwise what happens is that the main thread exits before the
processing thread drops its playbin, which causes the Paintable inside
the gtk4paintablesink to get dropped on the processing thread, as the
main thread's MainContext no longer exists.

We don't have to join on every window close, just right before exiting
the Application, but this is much simpler, and the join times aren't
overly bad usually.