page: Refactor prerolling to keep receiving bus messages
In some cases (like AV1), errors about missing codecs may arrive on
the playbin bus some time after the state change to Paused. In order to
catch them, we should keep receiving messages on the playbin bus after
prerolling.

Additionally, this commit changes playbin to be stored in the Page as
soon as it's created, and removed from the Page immediately on error.

Since adding a playbin to a Player's pipeline redirects all new messages
to the pipeline bus, we still need the Player logic to find and remove
an erroring playbin.